@@ -2,22 +2,24 @@
   "author": "devJimmyboy",
   "browserslist": "last 2 chrome versions",
   "dependencies": {
-    "@chakra-ui/react": "^1.8.6",
-    "postcss": "^8.4.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@chakra-ui/react": "^1.8.8",
+    "postcss": "^8.4.12",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "tailwindcss": "^3.0.23"
   },
   "description": "Template bundle for bundling with parcel.",
   "devDependencies": {
-    "@types/node": "^17.0.21",
-    "autoprefixer": "^10.4.2",
+    "@parcel/resolver-glob": "^2.4.1",
+    "@types/node": "^17.0.23",
+    "autoprefixer": "^10.4.4",
     "npm-run-all": "^4.1.5",
-    "parcel": "^2.3.2",
-    "postcss-import": "^14.0.2",
-    "postcss-nesting": "^10.1.3",
+    "parcel": "^2.4.1",
+    "postcss-import": "^14.1.0",
+    "postcss-nesting": "^10.1.4",
     "postcss-url": "^10.1.3",
-    "typescript": "^4.6.2"
+    "process": "^0.11.10",
+    "typescript": "^4.6.3"
   },
   "files": [
     "dashboard",
@@ -48,7 +50,7 @@
   },
   "scripts": {
     "build": "run-p build:*",
-    "build:dash": "parcel build --dist-dir dashboard --public-url ./",
+    "build:dash": "parcel build src/dashboard/panel.html --dist-dir dashboard --public-url ./",
     "build:ext": "tsc --project ./src/extension",
     "dev": "run-p dev:*",
     "dev:dash": "parcel watch src/dashboard/panel.html --dist-dir dashboard --public-url ./",
