@@ -3,35 +3,35 @@
     "chrome >= 95"
   ],
   "dependencies": {
-    "@eastdesire/jscolor": "^2.4.7",
-    "@iconify/iconify": "^2.1.2",
+    "@eastdesire/jscolor": "^2.4.8",
+    "@iconify/iconify": "^2.2.1",
     "animate.css": "^4.1.1",
-    "axios": "^0.26.0",
+    "axios": "^0.26.1",
     "bad-words": "^3.0.4",
-    "daisyui": "^2.6.0",
-    "gsap": "^3.9.1",
+    "daisyui": "^2.13.6",
+    "gsap": "^3.10.2",
     "jquery": "^3.6.0",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-select": "^5.2.2",
-    "use-nodecg": "^0.4.0"
+    "use-nodecg": "^0.4.1-1"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
-    "@types/lodash": "^4.14.179",
-    "@types/node": "^17.0.21",
-    "autoprefixer": "^10.4.2",
+    "@types/lodash": "^4.14.181",
+    "@types/node": "^17.0.23",
+    "autoprefixer": "^10.4.4",
     "nodecg-io-streamelements": "^0.2.0",
     "nodecg-types": "^1.8.3",
     "npm-run-all": "^4.1.5",
-    "parcel": "^2.3.2",
-    "postcss": "^8.4.7",
-    "postcss-import": "^14.0.2",
-    "postcss-nesting": "^10.1.3",
+    "parcel": "^2.4.1",
+    "postcss": "^8.4.12",
+    "postcss-import": "^14.1.0",
+    "postcss-nesting": "^10.1.4",
     "postcss-url": "^10.1.3",
     "tailwindcss": "^3.0.23",
-    "typescript": "^4.6.2"
+    "typescript": "^4.6.3"
   },
   "files": [
     "dashboard",
@@ -91,10 +91,10 @@
     ]
   },
   "scripts": {
-    "build": "run-p build:*",
-    "build:dashboard": "parcel build src/dashboard/*.html --dist-dir dashboard --public-url ./",
+    "build": "run-s build:*",
+    "build:dashboard": "parcel build src/dashboard/*.html --dist-dir ./dashboard --public-url ./",
     "build:extension": "tsc -b extension",
-    "build:graphics": "parcel build src/graphics/*.html --dist-dir graphics --public-url ./",
+    "build:graphics": "parcel build src/graphics/*.html --dist-dir ./graphics --public-url ./",
     "dev": "run-p dev:*",
     "dev:dashboard": "parcel watch src/dashboard/*.html --dist-dir ./dashboard --public-url . --port 1234",
     "dev:extension": "tsc -w --project extension/tsconfig.json",
