@@ -4,29 +4,28 @@
   ],
   "dependencies": {
     "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.6",
-    "@emotion/react": "^11.8.1",
+    "@chakra-ui/react": "^1.8.8",
+    "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@monaco-editor/react": "^4.3.1",
+    "@monaco-editor/react": "^4.4.1",
     "@twurple/api": "^5.0.18",
     "@twurple/auth": "^5.0.18",
     "@twurple/chat": "^5.0.18",
     "@twurple/eventsub": "^5.0.18",
     "@twurple/pubsub": "^5.0.18",
-    "@types/node": "^17.0.21",
+    "@types/node": "^17.0.23",
     "canvas-confetti": "^1.5.1",
-    "codemirror": "^5.65.2",
-    "daisyui": "^2.6.0",
+    "daisyui": "^2.13.6",
     "dlv": "^1.1.3",
     "dotenv": "^16.0.0",
-    "framer-motion": "^6.2.8",
-    "gsap": "^3.9.1",
+    "framer-motion": "^6.2.9",
+    "gsap": "^3.10.2",
     "lodash.debounce": "^4.0.8",
     "memoize-one": "^6.0.0",
-    "moment": "^2.29.1",
-    "monaco-editor": "0.32.1",
+    "moment": "^2.29.2",
+    "monaco-editor": "0.33.0",
     "nanotimer": "^0.3.15",
-    "nodecg": "^1.8.1",
+    "nodecg": "^1.9.0",
     "nodecg-io-core": "^0.2.0",
     "nodecg-io-streamelements": "^0.2.0",
     "nodecg-io-twitch-addons": "^0.2.0",
@@ -34,35 +33,35 @@
     "nodecg-io-twitch-chat": "^0.2.0",
     "nodecg-io-twitch-pubsub": "^0.2.0",
     "nodecg-types": "^1.8.3",
-    "postcss": "^8.4.7",
+    "postcss": "^8.4.12",
     "postcss-url": "^10.1.3",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-codemirror2": "^7.2.1",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.27.1",
+    "react-dom": "^18.0.0",
+    "react-hook-form": "^7.29.0",
     "react-select": "^5.2.2",
     "react-virtualized-auto-sizer": "^1.0.6",
-    "react-widgets": "^5.8.3",
+    "react-widgets": "^5.8.4",
     "react-window": "^1.8.6",
     "tailwindcss": "^3.0.23",
     "use-nodecg": "^0.4.1-1"
   },
   "devDependencies": {
-    "@parcel/config-default": "^2.3.2",
+    "@parcel/config-default": "^2.4.1",
     "@types/canvas-confetti": "^1.4.2",
     "@types/codemirror": "^5.60.5",
     "@types/dlv": "^1.1.2",
     "@types/lodash.debounce": "^4.0.6",
     "@types/nanotimer": "^0.3.0",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.13",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@types/react-virtualized-auto-sizer": "^1.0.1",
     "@types/react-window": "^1.8.5",
-    "@types/tailwindcss": "^3.0.9",
-    "autoprefixer": "^10.4.2",
+    "@types/tailwindcss": "^3.0.10",
+    "autoprefixer": "^10.4.4",
     "npm-run-all": "^4.1.5",
-    "postcss-import": "^14.0.2",
-    "typescript": "^4.6.2"
+    "postcss-import": "^14.1.0",
+    "typescript": "^4.6.3"
   },
   "name": "twitch",
   "nodecg": {
