@@ -1,17 +1,17 @@
-import React from "react"
-import ReactDOM from "react-dom"
-import { ChakraProvider, extendTheme } from "@chakra-ui/react"
+import { ChakraProvider, extendTheme } from "@chakra-ui/react";
+import React from "react";
+import { createRoot } from "react-dom/client";
 
-import App from "./App"
-const rootElement = document.getElementById("root")
-if (!rootElement) throw new Error("Failed to find the root element")
+import App from "./App";
+
+const rootElement = document.getElementById('root')
+if (!rootElement) throw new Error('Failed to find the root element')
 // const root = ReactDOM.createRoot(rootElement)
 
-const theme = extendTheme({ colors: require("daisyui/colors") })
+const theme = extendTheme()
 
-ReactDOM.render(
+createRoot(rootElement).render(
   <ChakraProvider theme={theme}>
     <App />
-  </ChakraProvider>,
-  rootElement
+  </ChakraProvider>
 )
