@@ -1,15 +1,10 @@
 module.exports = {
-  purge: [
-    "./src/dashboard/index.html",
-    "./src/graphics/index.html",
-    "./src/{dashboard,graphics}/*.{vue,js,ts,jsx,tsx,css}",
-  ],
-  darkMode: false, // or 'media' or 'class'
+  purge: ['./src/dashboard/index.html', './src/graphics/index.html', './src/{dashboard,graphics}/*.{vue,js,ts,jsx,tsx,css}'],
   theme: {
     extend: {},
   },
   variants: {
     extend: {},
   },
-  plugins: [],
+  plugins: [require('daisyui')],
 }
