@@ -1,31 +1,52 @@
-import React from 'react'
-import Select, { GroupBase, Options } from 'react-select'
-import { Alerts } from 'twitch/global'
-import { useReplicant } from 'use-nodecg'
+import React from "react";
+import Select from "react-select";
+import { Alerts } from "twitch/global";
+import { useReplicant } from "use-nodecg";
 
 interface SelectProps<Option = { value: number; label: string }, isMulti = true> {}
 
 type Props = { selectedAlert: Alerts.Alert | null; i: number; setAlert: (alert: Alerts.Alert, i: number) => void }
 
 export default function AlertsWrap({ selectedAlert, setAlert, i }: Props) {
-  const [media] = useReplicant<Alerts.Asset[]>('assets:media-graphics', [])
-  if (selectedAlert === null) return <div id="alerts-wrap"></div>
+  const [media] = useReplicant<Alerts.Asset[], Alerts.Asset[]>('assets:media-graphics', [])
+  if (!selectedAlert) return <div id="alerts-wrap"></div>
   else
     return (
-      <div id="alerts-wrap" className="flex flex-col items-center w-full">
+      <div id="alerts-wrap" className="flex flex-col items-center w-full gap-6">
         <div>
-          <label htmlFor="name">Name:</label>
-          <input id="name" type="text" value={selectedAlert.name} onChange={(e) => setAlert({ ...selectedAlert, name: e.target.value }, i)} />
+          <label htmlFor="name" className="mb-2 text-white">
+            Name:
+          </label>
+          <input
+            id="name"
+            className="input input-bordered text-base-content font-semibold"
+            type="text"
+            value={selectedAlert.name}
+            onChange={(e) => setAlert({ ...selectedAlert, name: e.target.value }, i)}
+          />
         </div>
         <div>
-          <label htmlFor="duration">Duration:</label>
-          <input id="duration" type="number" step={1} value={selectedAlert.duration} onChange={(e) => setAlert({ ...selectedAlert, duration: parseInt(e.target.value) }, i)} />
+          <label htmlFor="duration" className="mb-2 text-white">
+            Duration:
+          </label>
+          <input
+            id="duration"
+            className="input input-bordered text-base-content font-semibold"
+            type="number"
+            step={1}
+            value={selectedAlert.duration}
+            onChange={(e) => setAlert({ ...selectedAlert, duration: parseInt(e.target.value) }, i)}
+          />
         </div>
         <div className="text-gray-800 w-1/2">
+          <label htmlFor="media" className="mb-2 text-white">
+            Videos/Memes:
+          </label>
           <Select<{ value: number; label: string }, true>
             getOptionLabel={(option) => option.label}
             getOptionValue={(option) => option.value.toString()}
             isMulti
+            id="media"
             value={selectedAlert.media.map((m) => media[m] && { value: m, label: media[m].base })}
             className="text-gray-800 w-full"
             // @ts-ignore
