@@ -1,21 +1,22 @@
-import React from 'react'
-import { useEffect } from 'react'
-import Select, { ActionMeta, GroupBase } from 'react-select'
-import { Alerts } from 'twitch/global'
-import { useReplicant } from 'use-nodecg'
-import AlertsWrap from './AlertsWrap'
+import React from "react";
+import { useEffect } from "react";
+import Select, { ActionMeta, GroupBase } from "react-select";
+import { Alerts } from "twitch/global";
+import { useReplicant } from "use-nodecg";
+
+import AlertsWrap from "./AlertsWrap";
 
 type Props = {}
 
 interface SelectProps<Option = string, IsMulti extends boolean = false, Group extends GroupBase<Option> = GroupBase<Option>> {}
 
 export default function App({}: Props) {
-  const [alerts, setAlerts] = useReplicant<Alerts.Alert[]>('alerts', [])
-  const [assets, setAssets] = useReplicant<Alerts.Asset[]>('assets:media-graphics', [])
+  const [alerts, setAlerts] = useReplicant<Alerts.Alert[], Alerts.Alert[]>('alerts', [])
+  const [assets, setAssets] = useReplicant<Alerts.Asset[], Alerts.Asset[]>('assets:media-graphics', [])
   const [selectedAlert, setSelectedAlert] = React.useState<Alerts.Alert | null>(null)
   useEffect(() => {
     console.log(alerts)
-    if (selectedAlert !== null) setSelectedAlert(alerts.find((alert) => alert.name === selectedAlert.name))
+    if (selectedAlert) setSelectedAlert(alerts.find((alert) => alert.name === selectedAlert.name))
   }, [alerts])
 
   const onSelect = (option: { value: string; label: string }, actionMeta: ActionMeta<string>) => {
@@ -28,11 +29,12 @@ export default function App({}: Props) {
   const setAlert = function (alert: Alerts.Alert, i: number): void {
     const newAlerts = []
     newAlerts.push(...alerts.slice(0, i), alert, ...alerts.slice(i + 1))
+    setSelectedAlert(alert)
     setAlerts(newAlerts)
   }
 
   return (
-    <div className={`${selectedAlert === null ? 'mb-48' : 'mb-12'}`}>
+    <div className={`flex flex-col ${selectedAlert === null ? 'mb-48' : 'mb-12'} h-full`}>
       <div className="flex flex-row w-full justify-center items-end gap-2 pb-4">
         <div className="form-control w-full max-w-xs p-0 ">
           <label className="label">
@@ -57,7 +59,30 @@ export default function App({}: Props) {
           )}
         </div>
 
-        <button id="newAlert" className="btn btn-background self-end">
+        <button
+          id="newAlert"
+          className="btn btn-background self-end"
+          onClick={(e) =>
+            setAlerts(
+              alerts.concat([
+                {
+                  name: 'New Alert',
+                  customCSS: '',
+                  duration: 0,
+                  font: 'Lilita One',
+                  fontSize: '26',
+                  fontColour: '#ffffff',
+                  fontWeight: '',
+                  keywordColour: '#1261a4',
+                  message: '',
+                  layout: 'above',
+                  media: [],
+                  sound: '',
+                  volume: 0,
+                },
+              ])
+            )
+          }>
           <span className="iconify" data-icon="fa-solid:plus" style={{ fontSize: '24px', color: 'white' }}></span>
         </button>
         <button
@@ -87,6 +112,15 @@ export default function App({}: Props) {
         </button>
       </div>
       <AlertsWrap selectedAlert={selectedAlert} setAlert={setAlert} i={alerts.findIndex((a) => a.name === selectedAlert?.name)} />
+      <div className="flex flex-col w-full flex-grow items-center justify-center">
+        <div>types of events:</div>
+        <div>cheer</div>
+        <div>gift-subscriber</div>
+        <div>subscriber</div>
+        <div>raid</div>
+        <div>host</div>
+        <div>follow</div>
+      </div>
     </div>
   )
 }
